refactor(spree): extract checkout submission logic to separate utility
